requirements.txt
pandas==1.0.3
numpy==1.18.2
sympy==1.7.1
streamlit==0.71.0
Pillow==8.1.0
